Move app.js to the root folder and correct it accordingly
